+++ src/request/lifecycle.ts
@@ -31,8 +31,4 @@ import {type RequestPhase} from './phase';
  *
  * @category Requests
  */
-export class RequestLifecycle extends Lifecycle<RequestPhase> {
-	constructor() {
-		super();
-	}
-}
+export class RequestLifecycle extends Lifecycle<RequestPhase> {}